@@ -13,7 +13,7 @@
  ********************************************************************************/
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { webSocketURL } from '../../../configuration';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { isNullOrUndefined } from 'util';
