@@ -13,7 +13,7 @@
  *******************************************************************************/
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { backendBaseURL } from '../../configuration';
 import { HttpClient } from '@angular/common/http';
 
