@@ -13,7 +13,7 @@
  *******************************************************************************/
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { SelectItem } from 'ng2-select';
 import { WineryComponent } from '../../../../wineryInterfaces/wineryComponent';
 import { backendBaseURL } from '../../../../configuration';
