@@ -12,7 +12,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  ********************************************************************************/
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Rx';
+import { Observable } from 'rxjs';
 import { QNameWithTypeApiData } from '../wineryInterfaces/qNameWithTypeApiData';
 import { backendBaseURL } from '../configuration';
 import { InstanceService } from '../instance/instance.service';
