@@ -11,10 +11,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-import {WineryEditorComponent} from './wineryEditor.component';
-import {FormsModule} from '@angular/forms';
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { WineryEditorComponent } from './wineryEditor.component';
+import { FormsModule } from '@angular/forms';
+import { EditorBuilderService } from './editor-builder.service';
 
 @NgModule({
     imports: [
@@ -25,7 +26,7 @@ import {FormsModule} from '@angular/forms';
         WineryEditorComponent
     ],
     declarations: [WineryEditorComponent],
-    providers: [],
+    providers: [EditorBuilderService],
 })
 export class WineryEditorModule {
 }
